Rename files in gateway code
Since "local" gateway doesn't have the
local bridge anymore and its pretty much
"shared" gateway and both these gateways
are sharing the interface with the node,
we need to update our file naming
conventions to reflect code in a more
meaningful way even if we cannot really do
anything about the `gatewayConfig.mode`
naming scheme since we'd break compatibility.
