+++ go-controller/pkg/node/gateway_interface.go
@@ -10,7 +10,6 @@ import (
 
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/config"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/factory"
-	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/kube"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/types"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/util"
 
@@ -755,48 +754,6 @@ func (npwipt *nodePortWatcherIptables) SyncServices(services []interface{}) {
 	}
 }
 
-// since we share the host's k8s node IP, add OpenFlow flows
-// -- to steer the NodePort traffic arriving on the host to the OVN logical topology and
-// -- to also connection track the outbound north-south traffic through l3 gateway so that
-//    the return traffic can be steered back to OVN logical topology
-// -- to handle host -> service access, via masquerading from the host to OVN GR
-// -- to handle external -> service(ExternalTrafficPolicy: Local) -> host access without SNAT
-func newSharedGatewayOpenFlowManager(gwBridge, exGWBridge *bridgeConfiguration) (*openflowManager, error) {
-	dftFlows, err := flowsForDefaultBridge(gwBridge.ofPortPhys, gwBridge.macAddress.String(), gwBridge.ofPortPatch,
-		gwBridge.ofPortHost, gwBridge.ips)
-	if err != nil {
-		return nil, err
-	}
-	dftCommonFlows := commonFlows(gwBridge.ofPortPhys, gwBridge.macAddress.String(), gwBridge.ofPortPatch,
-		gwBridge.ofPortHost)
-	dftFlows = append(dftFlows, dftCommonFlows...)
-
-	// add health check function to check default OpenFlow flows are on the shared gateway bridge
-	ofm := &openflowManager{
-		defaultBridge:         gwBridge,
-		externalGatewayBridge: exGWBridge,
-		flowCache:             make(map[string][]string),
-		flowMutex:             sync.Mutex{},
-		exGWFlowCache:         make(map[string][]string),
-		exGWFlowMutex:         sync.Mutex{},
-		flowChan:              make(chan struct{}, 1),
-	}
-
-	ofm.updateFlowCacheEntry("NORMAL", []string{fmt.Sprintf("table=0,priority=0,actions=%s\n", util.NormalAction)})
-	ofm.updateFlowCacheEntry("DEFAULT", dftFlows)
-
-	// we consume ex gw bridge flows only if that is enabled
-	if exGWBridge != nil {
-		ofm.updateExBridgeFlowCacheEntry("NORMAL", []string{fmt.Sprintf("table=0,priority=0,actions=%s\n", util.NormalAction)})
-		exGWBridgeDftFlows := commonFlows(exGWBridge.ofPortPhys, exGWBridge.macAddress.String(),
-			exGWBridge.ofPortPatch, exGWBridge.ofPortHost)
-		ofm.updateExBridgeFlowCacheEntry("DEFAULT", exGWBridgeDftFlows)
-	}
-
-	// defer flowSync until syncService() to prevent the existing service OpenFlows being deleted
-	return ofm, nil
-}
-
 func flowsForDefaultBridge(ofPortPhys, bridgeMacAddress, ofPortPatch, ofPortHost string, bridgeIPs []*net.IPNet) ([]string, error) {
 	var dftFlows []string
 	// 14 bytes of overhead for ethernet header (does not include VLAN)
@@ -1163,81 +1120,6 @@ func setBridgeOfPorts(bridge *bridgeConfiguration) error {
 	return nil
 }
 
-func newSharedGateway(nodeName string, subnets []*net.IPNet, gwNextHops []net.IP, gwIntf, egressGWIntf string,
-	gwIPs []*net.IPNet, nodeAnnotator kube.Annotator, kube kube.Interface, cfg *managementPortConfig, watchFactory factory.NodeWatchFactory) (*gateway, error) {
-	klog.Info("Creating new shared gateway")
-	gw := &gateway{}
-
-	gwBridge, exGwBridge, err := gatewayInitInternal(
-		nodeName, gwIntf, egressGWIntf, subnets, gwNextHops, gwIPs, nodeAnnotator)
-	if err != nil {
-		return nil, err
-	}
-
-	if config.OvnKubeNode.Mode == types.NodeModeFull {
-		// add masquerade subnet route to avoid zeroconf routes
-		err = addMasqueradeRoute(gwBridge.bridgeName, gwNextHops)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	if exGwBridge != nil {
-		gw.readyFunc = func() (bool, error) {
-			ready, err := gatewayReady(gwBridge.patchPort)
-			if err != nil {
-				return false, err
-			}
-			exGWReady, err := gatewayReady(exGwBridge.patchPort)
-			if err != nil {
-				return false, err
-			}
-			return ready && exGWReady, nil
-		}
-	} else {
-		gw.readyFunc = func() (bool, error) {
-			return gatewayReady(gwBridge.patchPort)
-		}
-	}
-
-	gw.initFunc = func() error {
-		// Program cluster.GatewayIntf to let non-pod traffic to go to host
-		// stack
-		klog.Info("Creating Shared Gateway Openflow Manager")
-		err := setBridgeOfPorts(gwBridge)
-		if err != nil {
-			return err
-		}
-		if exGwBridge != nil {
-			err = setBridgeOfPorts(exGwBridge)
-			if err != nil {
-				return err
-			}
-		}
-		gw.openflowManager, err = newSharedGatewayOpenFlowManager(gwBridge, exGwBridge)
-		if err != nil {
-			return err
-		}
-
-		gw.nodeIPManager = newAddressManager(nodeName, kube, cfg, watchFactory)
-
-		if config.Gateway.NodeportEnable {
-			klog.Info("Creating Shared Gateway Node Port Watcher")
-			gw.nodePortWatcher, err = newNodePortWatcher(gwBridge.patchPort, gwBridge.bridgeName, gwBridge.uplinkName, gwBridge.ips, gw.openflowManager, gw.nodeIPManager, watchFactory)
-			if err != nil {
-				return err
-			}
-		} else {
-			// no service OpenFlows, request to sync flows now.
-			gw.openflowManager.requestFlowSync()
-		}
-		return nil
-	}
-
-	klog.Info("Shared Gateway Creation Complete")
-	return gw, nil
-}
-
 func newNodePortWatcher(patchPort, gwBridge, gwIntf string, ips []*net.IPNet, ofm *openflowManager,
 	nodeIPManager *addressManager, watchFactory factory.NodeWatchFactory) (*nodePortWatcher, error) {
 	// Get ofport of patchPort
@@ -1298,7 +1180,7 @@ func newNodePortWatcher(patchPort, gwBridge, gwIntf string, ips []*net.IPNet, of
 	return npw, nil
 }
 
-func cleanupSharedGateway() error {
+func cleanupGateway() error {
 	// NicToBridge() may be created before-hand, only delete the patch port here
 	stdout, stderr, err := util.RunOVSVsctl("--columns=name", "--no-heading", "find", "port",
 		"external_ids:ovn-localnet-port!=_")
@@ -1339,7 +1221,7 @@ func cleanupSharedGateway() error {
 		return fmt.Errorf("failed to replace-flows on bridge %q stderr:%s (%v)", bridgeName, stderr, err)
 	}
 
-	cleanupSharedGatewayIPTChains()
+	cleanupGatewayIPTChains()
 	return nil
 }
